+++ bindings/python/Math/tests/test_quintic_spline.py
@@ -1,5 +1,5 @@
 import pytest
-pytestmark = pytest.mark.planners
+pytestmark = pytest.mark.math
 
 import bipedal_locomotion_framework.bindings as blf
 import numpy as np
@@ -50,7 +50,7 @@ def test_quintic_spline():
         + 4 * 3 * coefficients[4] * (final_time ** 2) \
         + 5 * 4 * coefficients[5] * (final_time ** 3)
 
-    spline = blf.planners.QuinticSpline()
+    spline = blf.math.QuinticSpline()
 
     assert spline.set_knots(knots, time)
     assert spline.set_initial_conditions(init_velocity, init_acceleration)