+++ src/Math/tests/SplineTest.cpp
@@ -12,11 +12,10 @@
 // Catch2
 #include <catch2/catch_test_macros.hpp>
 
-#include <BipedalLocomotion/Planners/QuinticSpline.h>
-#include <BipedalLocomotion/Planners/CubicSpline.h>
+#include <BipedalLocomotion/Math/CubicSpline.h>
+#include <BipedalLocomotion/Math/QuinticSpline.h>
 
-using namespace BipedalLocomotion::Planners;
-
+using namespace BipedalLocomotion::Math;
 
 TEST_CASE("Quintic spline")
 {
@@ -28,52 +27,47 @@ TEST_CASE("Quintic spline")
 
     constexpr std::chrono::nanoseconds dT = (finalTime - initTime) / (numberOfPoints - 1);
     std::array<Eigen::Vector4d, 6> coefficients;
-    for(auto& coeff: coefficients)
+    for (auto& coeff : coefficients)
     {
         coeff.setRandom();
     }
 
-    std::vector<Eigen::VectorXd> knots;
+    std::vector<Eigen::Vector4d> knots;
     std::vector<std::chrono::nanoseconds> time;
     for (std::size_t i = 0; i < numberOfPoints; i++)
     {
         time.push_back(dT * i + initTime);
         const double t = std::chrono::duration<double>(time.back()).count();
-        knots.push_back(coefficients[0] + coefficients[1] * t
-                        + coefficients[2] * std::pow(t, 2)
-                        + coefficients[3] * std::pow(t, 3)
-                        + coefficients[4] * std::pow(t, 4)
+        knots.push_back(coefficients[0] + coefficients[1] * t + coefficients[2] * std::pow(t, 2)
+                        + coefficients[3] * std::pow(t, 3) + coefficients[4] * std::pow(t, 4)
                         + coefficients[5] * std::pow(t, 5));
     }
 
     const double initT = std::chrono::duration<double>(initTime).count();
     const double finalT = std::chrono::duration<double>(finalTime).count();
-    Eigen::Vector4d initVelocity = coefficients[1] + 2 * coefficients[2] * initT
-                                   + 3 * coefficients[3] * std::pow(initT, 2)
-                                   + 4 * coefficients[4] * std::pow(initT, 3)
-                                   + 5 * coefficients[5] * std::pow(initT, 4);
-
+    Eigen::Vector4d initVelocity
+        = coefficients[1] + 2 * coefficients[2] * initT + 3 * coefficients[3] * std::pow(initT, 2)
+          + 4 * coefficients[4] * std::pow(initT, 3) + 5 * coefficients[5] * std::pow(initT, 4);
 
     Eigen::Vector4d initAcceleration = 2 * coefficients[2] + 3 * 2 * coefficients[3] * initT
                                        + 4 * 3 * coefficients[4] * std::pow(initT, 2)
                                        + 5 * 4 * coefficients[5] * std::pow(initT, 3);
 
-    Eigen::Vector4d finalVelocity = coefficients[1] + 2 * coefficients[2] * finalT
-                                    + 3 * coefficients[3] * std::pow(finalT, 2)
-                                    + 4 * coefficients[4] * std::pow(finalT, 3)
-                                    + 5 * coefficients[5] * std::pow(finalT, 4);
+    Eigen::Vector4d finalVelocity
+        = coefficients[1] + 2 * coefficients[2] * finalT + 3 * coefficients[3] * std::pow(finalT, 2)
+          + 4 * coefficients[4] * std::pow(finalT, 3) + 5 * coefficients[5] * std::pow(finalT, 4);
 
     Eigen::Vector4d finalAcceleration = 2 * coefficients[2] + 3 * 2 * coefficients[3] * finalT
                                         + 4 * 3 * coefficients[4] * std::pow(finalT, 2)
                                         + 5 * 4 * coefficients[5] * std::pow(finalT, 3);
 
-    QuinticSpline spline;
+    QuinticSpline<Eigen::Vector4d> spline;
     REQUIRE(spline.setKnots(knots, time));
 
-    REQUIRE(spline.setInitialConditions(initVelocity, initAcceleration));
-    REQUIRE(spline.setFinalConditions(finalVelocity, finalAcceleration));
+    REQUIRE(spline.setInitialConditions({initVelocity, initAcceleration}));
+    REQUIRE(spline.setFinalConditions({finalVelocity, finalAcceleration}));
 
-    constexpr std::size_t pointsToCheckNumber = 1e4;
+    constexpr std::size_t pointsToCheckNumber = 1e3;
 
     constexpr std::chrono::nanoseconds dTCheckPoints
         = (finalTime - initTime) / (pointsToCheckNumber);
@@ -88,22 +82,26 @@ TEST_CASE("Quintic spline")
         REQUIRE(spline.evaluatePoint(t, position, velocity, acceleration));
 
         // check position
-        expected = coefficients[0] + coefficients[1] * tDouble + coefficients[2] * std::pow(tDouble, 2)
-                   + coefficients[3] * std::pow(tDouble, 3) + coefficients[4] * std::pow(tDouble, 4)
+        expected = coefficients[0] + coefficients[1] * tDouble
+                   + coefficients[2] * std::pow(tDouble, 2) + coefficients[3] * std::pow(tDouble, 3)
+                   + coefficients[4] * std::pow(tDouble, 4)
                    + coefficients[5] * std::pow(tDouble, 5);
 
         REQUIRE(expected.isApprox(position, 1e-5));
 
         // check velocity
-        expected = coefficients[1] + 2 * coefficients[2] * tDouble + 3 * coefficients[3] * std::pow(tDouble, 2)
-                   + 4 * coefficients[4] * std::pow(tDouble, 3) + 5 * coefficients[5] * std::pow(tDouble, 4);
-        REQUIRE(expected.isApprox(velocity,  1e-5));
+        expected = coefficients[1] + 2 * coefficients[2] * tDouble
+                   + 3 * coefficients[3] * std::pow(tDouble, 2)
+                   + 4 * coefficients[4] * std::pow(tDouble, 3)
+                   + 5 * coefficients[5] * std::pow(tDouble, 4);
+        REQUIRE(expected.isApprox(velocity, 1e-5));
 
         // check acceleration
         expected = 2 * coefficients[2] + 3 * 2 * coefficients[3] * tDouble
                    + 4 * 3 * coefficients[4] * std::pow(tDouble, 2)
                    + 5 * 4 * coefficients[5] * std::pow(tDouble, 3);
-        REQUIRE(expected.isApprox(acceleration,  1e-5));
+
+        REQUIRE(expected.isApprox(acceleration, 1e-5));
     }
 
     SECTION("Advance capabilities")
@@ -111,11 +109,13 @@ TEST_CASE("Quintic spline")
         REQUIRE_FALSE(spline.isOutputValid());
         REQUIRE(spline.setAdvanceTimeStep(dTCheckPoints));
 
-        REQUIRE(spline.isOutputValid());
-
         for (std::size_t i = 0; i < pointsToCheckNumber; i++)
         {
             double t = std::chrono::duration<double>(dTCheckPoints * i + initTime).count();
+
+            // advance the spline
+            REQUIRE(spline.advance());
+            REQUIRE(spline.isOutputValid());
             const auto& traj = spline.getOutput();
 
             // check position
@@ -136,13 +136,48 @@ TEST_CASE("Quintic spline")
                        + 4 * 3 * coefficients[4] * std::pow(t, 2)
                        + 5 * 4 * coefficients[5] * std::pow(t, 3);
             REQUIRE(expected.isApprox(traj.acceleration, 1e-5));
+        }
+    }
 
-            // advance the spline
-            REQUIRE(spline.advance());
+    SECTION("Query from a vector of times")
+    {
+        std::vector<std::chrono::nanoseconds> timeVector;
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            timeVector.push_back(dTCheckPoints * i + initTime);
+        }
+
+        std::vector<Eigen::Vector4d> positionVector, velocityVector, accelerationVector;
+        REQUIRE(spline.evaluateOrderedPoints(timeVector, //
+                                             positionVector,
+                                             velocityVector,
+                                             accelerationVector));
+
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            double t = std::chrono::duration<double>(dTCheckPoints * i + initTime).count();
+
+            // check position
+            expected = coefficients[0] + coefficients[1] * t + coefficients[2] * std::pow(t, 2)
+                       + coefficients[3] * std::pow(t, 3) + coefficients[4] * std::pow(t, 4)
+                       + coefficients[5] * std::pow(t, 5);
+
+            REQUIRE(expected.isApprox(positionVector[i], 1e-5));
+
+            // check velocity
+            expected = coefficients[1] + 2 * coefficients[2] * t
+                       + 3 * coefficients[3] * std::pow(t, 2) + 4 * coefficients[4] * std::pow(t, 3)
+                       + 5 * coefficients[5] * std::pow(t, 4);
+            REQUIRE(expected.isApprox(velocityVector[i], 1e-5));
+
+            // check acceleration
+            expected = 2 * coefficients[2] + 3 * 2 * coefficients[3] * t
+                       + 4 * 3 * coefficients[4] * std::pow(t, 2)
+                       + 5 * 4 * coefficients[5] * std::pow(t, 3);
+            REQUIRE(expected.isApprox(accelerationVector[i], 1e-5));
         }
     }
 }
-
 
 TEST_CASE("Cubic spline")
 {
@@ -154,38 +189,40 @@ TEST_CASE("Cubic spline")
 
     constexpr std::chrono::nanoseconds dT = (finalTime - initTime) / (numberOfPoints - 1);
     std::array<Eigen::Vector4d, 6> coefficients;
-    for(auto& coeff: coefficients)
+    for (auto& coeff : coefficients)
     {
         coeff.setRandom();
     }
 
-    std::vector<Eigen::VectorXd> knots;
+    std::vector<Eigen::Vector4d> knots;
     std::vector<std::chrono::nanoseconds> time;
     for (std::size_t i = 0; i < numberOfPoints; i++)
     {
         time.push_back(dT * i + initTime);
         const double t = std::chrono::duration<double>(time.back()).count();
-        knots.push_back(coefficients[0] + coefficients[1] * t
-                        + coefficients[2] * std::pow(t, 2)
+        knots.push_back(coefficients[0] + coefficients[1] * t + coefficients[2] * std::pow(t, 2)
                         + coefficients[3] * std::pow(t, 3));
     }
 
     const double initT = std::chrono::duration<double>(initTime).count();
     const double finalT = std::chrono::duration<double>(finalTime).count();
 
-    Eigen::Vector4d initVelocity = coefficients[1] + 2 * coefficients[2] * initT
-                                   + 3 * coefficients[3] * std::pow(initT, 2);
+    Eigen::Vector4d initVelocity
+        = coefficients[1] + 2 * coefficients[2] * initT + 3 * coefficients[3] * std::pow(initT, 2);
 
     Eigen::Vector4d finalVelocity = coefficients[1] + 2 * coefficients[2] * finalT
                                     + 3 * coefficients[3] * std::pow(finalT, 2);
 
-    CubicSpline spline;
+    Eigen::Vector4d initAcceleration = 2 * coefficients[2] + 3 * 2 * coefficients[3] * initT;
+    Eigen::Vector4d finalAcceleration = 2 * coefficients[2] + 3 * 2 * coefficients[3] * finalT;
+
+    CubicSpline<Eigen::Vector4d> spline;
     REQUIRE(spline.setKnots(knots, time));
 
-    REQUIRE(spline.setInitialConditions(initVelocity));
-    REQUIRE(spline.setFinalConditions(finalVelocity));
+    REQUIRE(spline.setInitialConditions({initVelocity, initAcceleration}));
+    REQUIRE(spline.setFinalConditions({finalVelocity, finalAcceleration}));
 
-    constexpr std::size_t pointsToCheckNumber = 1e4;
+    constexpr std::size_t pointsToCheckNumber = 1e3;
 
     constexpr std::chrono::nanoseconds dTCheckPoints = (finalTime - initTime) / pointsToCheckNumber;
 
@@ -208,8 +245,7 @@ TEST_CASE("Cubic spline")
         // check velocity
         expected = coefficients[1] + 2 * coefficients[2] * tSeconds
                    + 3 * coefficients[3] * std::pow(tSeconds, 2);
-        REQUIRE(expected.isApprox(velocity,  1e-5));
-
+        REQUIRE(expected.isApprox(velocity, 1e-5));
     }
 
     SECTION("Advance capabilities")
@@ -217,27 +253,60 @@ TEST_CASE("Cubic spline")
         REQUIRE_FALSE(spline.isOutputValid());
         REQUIRE(spline.setAdvanceTimeStep(dTCheckPoints));
 
-        REQUIRE(spline.isOutputValid());
-
         for (std::size_t i = 0; i < pointsToCheckNumber; i++)
         {
             double t = std::chrono::duration<double>(dTCheckPoints * i + initTime).count();
+
+            // advance the spline
+            REQUIRE(spline.advance());
+            REQUIRE(spline.isOutputValid());
             const auto& traj = spline.getOutput();
 
             // check position
-            expected = coefficients[0] + coefficients[1] * t + coefficients[2] * std::pow(t, 2)
-                + coefficients[3] * std::pow(t, 3);
+            expected = coefficients[0] //
+                       + coefficients[1] * t //
+                       + coefficients[2] * std::pow(t, 2) //
+                       + coefficients[3] * std::pow(t, 3);
 
             REQUIRE(expected.isApprox(traj.position, 1e-5));
 
             // check velocity
-            expected = coefficients[1] + 2 * coefficients[2] * t
-                + 3 * coefficients[3] * std::pow(t, 2);
+            expected = coefficients[1] //
+                       + 2 * coefficients[2] * t //
+                       + 3 * coefficients[3] * std::pow(t, 2);
             REQUIRE(expected.isApprox(traj.velocity, 1e-5));
+        }
+    }
 
+    SECTION("Query from a vector of times")
+    {
+        std::vector<std::chrono::nanoseconds> timeVector;
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            timeVector.push_back(dTCheckPoints * i + initTime);
+        }
 
-            // advance the spline
-            REQUIRE(spline.advance());
+        std::vector<Eigen::Vector4d> positionVector, velocityVector, accelerationVector;
+        REQUIRE(spline.evaluateOrderedPoints(timeVector, //
+                                             positionVector,
+                                             velocityVector,
+                                             accelerationVector));
+
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            double t = std::chrono::duration<double>(dTCheckPoints * i + initTime).count();
+
+            // check position
+            expected = coefficients[0] + coefficients[1] * t + coefficients[2] * std::pow(t, 2)
+                       + coefficients[3] * std::pow(t, 3);
+
+            REQUIRE(expected.isApprox(positionVector[i], 1e-5));
+
+            // check velocity
+            expected = coefficients[1] //
+                       + 2 * coefficients[2] * t //
+                       + 3 * coefficients[3] * std::pow(t, 2);
+            REQUIRE(expected.isApprox(velocityVector[i], 1e-5));
         }
     }
 }