+++ utilities/joint-trajectory-player/include/BipedalLocomotion/JointTrajectoryPlayer/Module.h
@@ -20,7 +20,6 @@
 #include <matioCpp/matioCpp.h>
 
 #include <BipedalLocomotion/ParametersHandler/IParametersHandler.h>
-#include <BipedalLocomotion/Planners/QuinticSpline.h>
 #include <BipedalLocomotion/RobotInterface/YarpHelper.h>
 #include <BipedalLocomotion/RobotInterface/YarpRobotControl.h>
 #include <BipedalLocomotion/RobotInterface/YarpSensorBridge.h>